style: slightly more accessible colors
The olive green and orange-red may be a bit more muted.
I tinkered a bit with the LCH color space to find some with the same
lightness while being accessible enough.

Contrast for "darkolivegreen": 5.9:1
Contrast for #B82F00:          6.1:1

Both should suffice for WCAG-G18.
I need to keep looking for something more colorful while
aiming for 7:1 color contrast.
This may be harder when designing a dark mode.