style: play around some more with gradient text
The accent-1 color iiisss somewhere between AA and AAA contrast
points of WCAG2 techniques G17 and G18. I heard once that when using
gradients around it's better to use the lighter of the ends to
determine if it should be readable. Somehow this almost makes it worse
and less readable than solid darkolivegreen.

Eh. Looks cool. Might change it later though.